requirement.txt
torch==1.13.1
torchvision==0.14.1
pandas==1.5.2
opencv-python==4.6.0.66
Pillow==9.4.0
scikit-learn==1.2.2
matplotlib==3.7.0
PyYAML==6.0
tqdm==4.64.1
scipy==1.10.1
